Set up ListingSidebar in discussions idx
